Adding save-by-id, it will update an existing document by that ID, or create one if the id is nil and/or is non-existing.
It is conflicting a bit conceptually with update-by-id, but update-by-id is meant for internal non-user-input atomic changes to documents, like $inc, $push etc.

No longer manually creating an ObjectId when inserting a new record, refraining from specifying an :_id at all is enough to make save-and-return create a new record. This was also true for the old insert-and-return, not sure why we actually created an ObjectId manually in the first place.