Removing validation completely from this library.
Validation always happens before everything else, and is not really related to mongodb operations. So when we used to do:

    (def insert (make-insert "my-collection" my-validator))

An equivalent would now be

    (def inserter (make-insert "my-collection"))
    (defn insert
      [data]
      (if-let [errors (my-validator data)]
        (if errors
          [false errors]
          [true (inserter data)])))

This is obviously more verbose, but code will be written (probably in oiiku-data-validations) to make this stuff easy.

This grants maximum composability and flexibility. The actual use case that made us discover this, is passing more than just the data to validators. With the old API that was not possible. Now that we compose stuff ourselves, it becomes trivial.